Calculodivi.java

+++ Practica 56/Reto/Calculodivi.java
@@ -1,13 +1,13 @@
 public class Calculodivi {
-     int num1;
-    int num2;
-    public Calculodivi(int num1, int num2){
+    double num1;
+    double num2;
+    public Calculodivi(double num1, double num2){
         this.num1=num1;
         this.num2=num2;
     }
-    public int obtenerDivi()
+    public double obtenerDivi()
     {
         return num1 / num2;
     }
-    //aqui va el proceso de la suma que ya me dio hueva escribir
+    
 }
